pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Mango", email = "evan940415@gmail.com"},
]
dependencies = [
    "py-cord>=2.1.3",
    "python-dotenv>=0.21.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
